+++ packages/app/src/components/modal.tsx
@@ -6,7 +6,7 @@ import {
   useId,
   useInteractions,
 } from "@floating-ui/react";
-import { TransitionV2 as Transition } from "@hiogawa/tiny-transition/dist/react";
+import { Transition } from "@hiogawa/tiny-transition/dist/react";
 import { tinyassert } from "@hiogawa/utils";
 import React from "react";
 import { cls } from "../utils/misc";
