+++ packages/tiny-transition/src/class.ts
@@ -1,8 +1,6 @@
 import type { TransitionCallbackProps } from "./core";
 
-// helper to use css class as a callback
-// TODO: but relying on callback means it won't be rendered during SSR.
-//       for example, when `show=true, appear=false`, we should render `entered` class during SSR.
+// helper to use css class for callbacks
 
 export const TRANSITION_CLASS_TYPES = [
   "enter",
