+++ packages/tiny-transition/src/index.ts
@@ -1,3 +1,2 @@
 export * from "./core";
 export * from "./class";
-export * from "./core-v2";
