+++ packages/tiny-transition/src/react.tsx
@@ -1,6 +1,6 @@
-import { objectOmit } from "@hiogawa/utils";
-import { useMergeRefs } from "@hiogawa/utils-react";
-import React from "react";
+import { objectMapValues, objectOmit } from "@hiogawa/utils";
+import { useMergeRefs, useStableCallback } from "@hiogawa/utils-react";
+import * as React from "react";
 import {
   TRANSITION_CLASS_TYPES,
   type TransitionClassProps,
@@ -11,81 +11,90 @@ import {
   type TransitionCallbackProps,
   TransitionManager,
 } from "./core";
-export * from "./react-v2";
 
-// cheat typing to simplify dts rollup
-function simpleForawrdRef<
-  P,
-  T,
-  F extends (props: P, ref: React.ForwardedRef<T>) => JSX.Element
->(f: F): (props: P & { ref?: React.Ref<T> }) => JSX.Element {
-  return React.forwardRef(f) as any;
+export function useTransitionManager(
+  value: boolean,
+  options?: { appear?: boolean; callbacks?: TransitionCallbackProps }
+) {
+  const [manager] = React.useState(
+    () =>
+      new TransitionManager(options?.appear ? false : value, options?.callbacks)
+  );
+
+  React.useEffect(() => {
+    manager.set(value);
+  }, [value]);
+
+  React.useSyncExternalStore(
+    manager.subscribe,
+    () => manager.state,
+    () => manager.state
+  );
+
+  return manager;
 }
 
 export const Transition = simpleForawrdRef(function Transition(
+  // prettier-ignore
   props: {
     show: boolean;
     appear?: boolean;
     render?: (props: Record<string, any>) => React.ReactNode;
-  } & TransitionClassProps &
-    TransitionCallbackProps & {
-      // choose only common props from `JSX.IntrinsicElements["div"]` to simplify auto-complete
-      className?: string;
-      style?: React.CSSProperties;
-      children?: React.ReactNode;
-    },
+    children?: React.ReactNode;
+    // all props will be delegated to <div /> but we pick only two to reduce noisy auto-completion
+    className?: string;
+    style?: React.CSSProperties;
+  } & TransitionClassProps
+    & TransitionCallbackProps,
   ref: React.ForwardedRef<HTMLElement>
 ) {
-  //
-  // setup TransitionManager as external store
-  //
-  const [manager] = React.useState(
-    () =>
-      new TransitionManager({
-        defaultEntered: Boolean(props.show && !props.appear),
-        ...convertClassPropsToCallbackProps(props.className, props),
+  // define stable callbacks with ref element
+  const elRef = React.useRef<HTMLElement | null>(null);
+  const callbacks = objectMapValues(
+    convertClassPropsToCallbackProps(props.className, props),
+    (callback) =>
+      useStableCallback(() => {
+        if (callback && elRef.current) {
+          callback(elRef.current);
+        }
       })
   );
-  const shouldRender = React.useSyncExternalStore(
-    manager.subscribe,
-    manager.shouldRender,
-    manager.shouldRender
-  );
 
-  //
-  // sync props to state
-  //
-  React.useEffect(() => {
-    manager.show(props.show ?? false);
-  }, [props.show]);
+  // transition manager
+  const manager = useTransitionManager(props.show, {
+    appear: props.appear,
+    callbacks,
+  });
 
-  React.useEffect(() => {
-    Object.assign(
-      manager.options,
-      convertClassPropsToCallbackProps(props.className, props)
-    );
-  }, [props]);
-
-  //
   // render
-  //
-  const delegatedProps = objectOmit(props, [
-    "show",
-    "appear",
-    "render",
-    ...TRANSITION_CLASS_TYPES,
-    ...TRANSITION_CALLBACK_TYPES,
-  ]);
-  const mergedRefs = useMergeRefs(ref, manager.setElement);
-  const render = props.render ?? ((props: any) => <div {...props} />);
+  const mergedRefs = useMergeRefs(ref, elRef, manager.ref);
+  const render = props.render ?? defaultRender;
 
+  // TODO: render `props.entered/enterTo` class when `show=true` and `appear=false`
   return (
-    <>
-      {shouldRender &&
-        render({
-          ref: mergedRefs,
-          ...delegatedProps,
-        })}
-    </>
+    manager.state &&
+    render({
+      ref: mergedRefs,
+      ...objectOmit(props, [
+        "show",
+        "appear",
+        "render",
+        ...TRANSITION_CLASS_TYPES,
+        ...TRANSITION_CALLBACK_TYPES,
+      ]),
+    })
   );
 });
+
+function defaultRender(props: any) {
+  return React.createElement("div", props);
+}
+
+// simplify forwardRef typing for dts rollup
+function simpleForawrdRef<
+  P,
+  T,
+  F extends (props: P, ref: React.ForwardedRef<T>) => React.ReactNode
+>(f: F): (props: P & { ref?: React.Ref<T> }) => React.ReactNode {
+  return React.forwardRef(f) as any;
+}
